requirements.txt
Django==5.0.6
djangorestframework==3.15.1
channels==4.1.0
channels-redis==4.2.0
redis==5.0.4
pymongo==4.7.2
neo4j==5.19.0
python-dotenv==1.0.1
uvicorn==0.29.0
gunicorn==22.0.0
drf-spectacular==0.27.2
django-redis==5.4.0
fakeredis==2.23.2
mongomock==4.1.2
pytest==8.2.2
pytest-django==4.8.0
Faker==25.2.0
